requirements.txt
appdirs==1.4.4
Django==5.0.7
djangorestframework==3.15.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
psycopg2-binary==2.9.9
django-crispy-forms==2.2
crispy-bootstrap5